requirements.txt
# App server
gunicorn==19.4.5

# Django
Django>=1.9,<2.0

# Ethereum
populus==0.7.3

# Alarm
ethereum-alarm-clock-client==0.7.0

# Ethereum
populus==0.7.3

# Django Storage
django-storages>=1.4
django-pipeline>=1.6.5
django-s3-folder-storage>=0.2
boto>=2.34.0

# REST
djangorestframework==3.3.2

# Images
Pillow>=3.1.1

# Database
psycopg2>=2.5.4

# Timezone support
pytz>=2014.10

# Configuration
python-dotenv>=0.1.2
env-excavator>=1.5.0
dj-database-url>=0.4.0
django-cache-url>=0.8.0

# Monitoring
rollbar>=0.11.2
PyReact==0.6.0
